Issue-20 Automatically detect the correct unit system from the FlowTracker2 file
There is no need for any user choice or configuration of unit systems, since the setting is already stored within the FlowTracker2 file.

1) 3.x converter

Remove the unit-selector combobox and use the file settings instead.

2) NG plugin

No need to look at a special file. Just read the file's units setting and use that.