Issue-34 - Calculate the measured depth when possible
The MeasD column was not being calculated correctly (at all).

The DIS report format is a bit odd.

Reverse engineering the calculations from the 3.X parser shows that:

MeasD = EffectiveDepth - Dep% * EffectiveDepth.

To make matters worse, multi-point verticals (eg. with two 80% and 20% observations) are never exported to DIS.

This change makes the the import work for 50% and 60% single-point observations, and for others, it just gives up.